rename dir shops to shop
