+++ admin-dev/themes/new-theme/.webpack/common.js
@@ -118,8 +118,8 @@ module.exports = {
     product_index: './js/pages/product/grid/index',
     product_page: './js/product-page/index',
     product_preferences: './js/pages/product-preferences',
-    product_shops: './js/pages/product/shops',
-    pre_select_product_shop: './js/pages/product/shops/pre-select-product-shop',
+    product_shops: './js/pages/product/shop',
+    pre_select_product_shop: './js/pages/product/shop/pre-select-product-shop',
     profiles: './js/pages/profiles',
     search_engine: './js/pages/search-engine',
     security: './js/pages/security',
