+++ admin-dev/themes/new-theme/js/pages/product/shop/index.ts
