+++ admin-dev/themes/new-theme/js/pages/product/shop/pre-select-product-shop/index.ts
@@ -24,7 +24,7 @@
  */
 
 import ProductMap from '@pages/product/product-map';
-import selectShopForEdition from '@pages/product/shops/select-shop-modal';
+import selectShopForEdition from '@pages/product/shop/select-shop-modal';
 
 document.addEventListener('DOMContentLoaded', () => {
   const warning = document.querySelector<HTMLElement>(ProductMap.shops.contextWarning);
