+++ admin-dev/themes/new-theme/js/pages/product/shop/product-shops-modal.ts
